requirements.txt
absl-py
tensorflow-gpu==2.0.0b1
tfds-nightly
numpy
tqdm